Move pretty into lib, because it's heavily changed now.
