+++ IPython/lib/pretty.py
